refactor: restructure docker/ into transparent/explicit/shared layout
