+++ docker/tools/shared/lib/rules.js
