+++ docker/tools/shared/lib/rules.test.js
