+++ docker/transparent/files/haproxy.cfg.template
