+++ docker/transparent/files/s6-rc.d/buildkitd/type
