+++ docker/transparent/files/s6-rc.d/dnsmasq/data/check
