+++ docker/transparent/files/s6-rc.d/haproxy/type
