+++ docker/transparent/files/s6-rc.d/init-haproxy-cfg/type
