+++ docker/transparent/files/s6-rc.d/init-iptables/type
